Use _check_broadcast_shapes and _check_all_safe_to_cast for uniform

+++ jax/experimental/rnn.py
@@ -171,7 +171,7 @@ def init_lstm_weight(rng: PRNGKeyArray, input_size: int, hidden_size: int,
   """Random initialize LSTM weights from U(-k, k), k=sqrt(1/hidden_size)."""
   param_count = get_num_params_in_lstm(input_size, hidden_size, num_layers,
                                        bidirectional)
-  k = np.sqrt(1.0 / hidden_size)
+  k = np.sqrt(1.0 / hidden_size).astype(jnp.float32)
   return jax.random.uniform(
       rng, shape=(param_count,), dtype=jnp.float32, minval=-k, maxval=k)
